third_party/WebKit/LayoutTests/fast/frames/seamless/seamless-hyperlink-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS frames[0].document.body.textContent.trim() is "DONE"
